refactor(cyclonedx): Factor out `CycloneDxModelMapper`
Extract a class dedicated to mapping between ORT's model and
CycloneDX's BOM model. This prepares for moving all model mapping
of this reporter to this new dedicated place in an upcoming change.

Note: The mapping input is provided to the constructor, to allow the use
      of caching to avoid re-computations, like for resolved license
      infos.
